requirements.txt
pandas
numpy
python-phonenumbers
nameparser
fuzzywuzzy
python-Levenshtein
scikit-learn
streamlit
plotly
matplotlib
seaborn
ipywidgets
notebook